Reorganize testing assets
Trigger CI
